Move repr stuff to utils module
